refactor :: Background 컴포넌트 구조 개선 및 절대 경로 적용
