+++ src/pages/auth/EnterPassword.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import PasswordBackground from '@shared/ui/Background/PasswordBackground';
+import { PasswordBackground } from '@shared/ui/Background/index';
 import LeftPage from '@widgets/layout/LeftPage';
 import RightPage from '@widgets/layout/RightPage';
 import { LoginForm } from '@features/auth/login/ui/LoginForm';
@@ -11,7 +11,6 @@ const EnterPassword = () => {
       <MainWrapper>
         <PageContainer>
           <LeftPage />
-
           <RightPage title="관리자 비밀번호 입력">
             <LoginForm />
           </RightPage>
