+++ src/pages/visit/UserVisitList.jsx
@@ -1,7 +1,7 @@
 import { useState, useRef, useEffect, useMemo } from 'react';
 import styled from '@emotion/styled';
 import { FaExclamationTriangle, FaCheckCircle } from 'react-icons/fa';
-import UseBackground from '@shared/ui/Background/UseBackground';
+import { UseBackground } from '@shared/ui/Background/index';
 import Header from '@widgets/layout/header/Header';
 import UserVisitCard from '@shared/ui/Form/UserVisitCard';
 import ExcelButton from '@shared/ui/Button/ExcelButton';
