+++ src/shared/ui/Background/index.ts
@@ -1,0 +1,3 @@
+export { CircleData } from './CircleData';
+export { PasswordBackground } from './PasswordBackground';
+export { UseBackground } from './UseBackground';
